Consolidate in the Geometry_Engine
